리팩토링: My...Item, My...List로 컴포넌트 분리
